Move clientID to header utils
