+++ pkg/utils/client_id.go
@@ -1,10 +1,13 @@
-package server
+package utils
 
 import (
 	"context"
 
-	"github.com/mikekulinski/zookeeper/pkg/client"
 	"google.golang.org/grpc/metadata"
+)
+
+const (
+	ClientIDHeader = "X-Client-ID"
 )
 
 // ExtractClientIDHeader extracts the clientID from the context.
@@ -14,10 +17,17 @@ func ExtractClientIDHeader(ctx context.Context) (string, bool) {
 		return "", false
 	}
 
-	values := md.Get(client.ClientIDHeader)
+	values := md.Get(ClientIDHeader)
 	if len(values) == 0 {
 		return "", false
 	}
 
 	return values[0], true
+}
+
+func SetClientIDHeader(ctx context.Context, clientID string) context.Context {
+	// Add client ID to outgoing metadata
+	md := metadata.Pairs(ClientIDHeader, clientID)
+	ctx = metadata.NewOutgoingContext(ctx, md)
+	return ctx
 }
